Move files around, modify gl init, cleanups, magic--
