+++ libdragoncast/deps/khronos/EGL/egl.h
@@ -33,7 +33,7 @@
 /* All platform-dependent types and macro boilerplate (such as EGLAPI
  * and EGLAPIENTRY) should go in eglplatform.h.
  */
-#include <EGL/eglplatform.h>
+#include "../EGL/eglplatform.h"
 
 #ifdef __cplusplus
 extern "C" {
