+++ libdragoncast/deps/khronos/GL4/glcorearb.h
@@ -79,7 +79,7 @@ extern "C" {
 #define GL_VERSION_1_0 1
 typedef void GLvoid;
 typedef unsigned int GLenum;
-#include <KHR/khrplatform.h>
+#include "../KHR/khrplatform.h"
 typedef khronos_float_t GLfloat;
 typedef int GLint;
 typedef int GLsizei;
