+++ libdragoncast/glwrap/gl3w.cpp
@@ -26,7 +26,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include <GL4/gl3w.h>
+#include "gl3w.h"
 #include <stdlib.h>
 
 #define ARRAY_SIZE(x)  (sizeof(x) / sizeof((x)[0]))
